+++ tests/unit/oid4vp.spec.js
@@ -1,5 +1,5 @@
 /*!
- * Copyright (c) 2022-2023 Digital Bazaar, Inc. All rights reserved.
+ * Copyright (c) 2022-2024 Digital Bazaar, Inc. All rights reserved.
  */
 import {_fromQueryByExampleQuery} from '../../lib/oid4vp.js';
 import chai from 'chai';
@@ -8,8 +8,32 @@ chai.should();
 const {expect} = chai;
 
 describe('OID4VP', () => {
-  describe('constructor', () => {
-    it('should map a QueryByExample to a Presentation Definition', async () => {
+  describe('QueryByExample => Presentation Definition', () => {
+    it('should NOT include "vc" prefix in paths', async () => {
+      const presentation_definition = _fromQueryByExampleQuery({
+        credentialQuery: {
+          reason: 'Please present your Driver\'s License to complete the ' +
+            'verification process.',
+          example: {
+            '@context': [
+              'https://www.w3.org/2018/credentials/v1',
+              'https://w3id.org/vdl/v1',
+              'https://w3id.org/vdl/aamva/v1'
+            ],
+            type: [
+              'Iso18013DriversLicenseCredential'
+            ]
+          }
+        },
+        prefixJwtVcPath: false
+      });
+      expect(presentation_definition.constraints.fields[0].path).to.eql(
+        ['$[\'@context\']']);
+      expect(presentation_definition.constraints.fields[1].path).to.eql(
+        ['$[\'type\']']);
+    });
+
+    it('should include "vc" prefix in paths', async () => {
       const presentation_definition = _fromQueryByExampleQuery({
         credentialQuery: {
           reason: 'Please present your Driver\'s License to complete the ' +
@@ -26,6 +50,81 @@ describe('OID4VP', () => {
           }
         },
         prefixJwtVcPath: true
+      });
+      expect(presentation_definition.constraints.fields[0].path).to.eql([
+        '$[\'@context\']',
+        '$[\'vc\'][\'@context\']'
+      ]);
+      expect(presentation_definition.constraints.fields[1].path).to.eql([
+        '$[\'type\']',
+        '$[\'vc\'][\'type\']'
+      ]);
+    });
+
+    it('auto-detect to NOT include "vc" w/unspecified security', async () => {
+      const presentation_definition = _fromQueryByExampleQuery({
+        credentialQuery: {
+          reason: 'Please present your Driver\'s License to complete the ' +
+            'verification process.',
+          example: {
+            '@context': [
+              'https://www.w3.org/2018/credentials/v1',
+              'https://w3id.org/vdl/v1',
+              'https://w3id.org/vdl/aamva/v1'
+            ],
+            type: [
+              'Iso18013DriversLicenseCredential'
+            ]
+          }
+        }
+      });
+      expect(presentation_definition.constraints.fields[0].path).to.eql(
+        ['$[\'@context\']']);
+      expect(presentation_definition.constraints.fields[1].path).to.eql(
+        ['$[\'type\']']);
+    });
+
+    it('auto-detect to NOT include "vc" w/acceptedCryptosuites', async () => {
+      const presentation_definition = _fromQueryByExampleQuery({
+        credentialQuery: {
+          reason: 'Please present your Driver\'s License to complete the ' +
+            'verification process.',
+          example: {
+            '@context': [
+              'https://www.w3.org/2018/credentials/v1',
+              'https://w3id.org/vdl/v1',
+              'https://w3id.org/vdl/aamva/v1'
+            ],
+            type: [
+              'Iso18013DriversLicenseCredential'
+            ]
+          },
+          acceptedCryptosuites: ['bbs-2023']
+        }
+      });
+      expect(presentation_definition.constraints.fields[0].path).to.eql(
+        ['$[\'@context\']']);
+      expect(presentation_definition.constraints.fields[1].path).to.eql(
+        ['$[\'type\']']);
+    });
+
+    it('auto-detect to include "vc" w/acceptedEnvelopes', async () => {
+      const presentation_definition = _fromQueryByExampleQuery({
+        credentialQuery: {
+          reason: 'Please present your Driver\'s License to complete the ' +
+            'verification process.',
+          example: {
+            '@context': [
+              'https://www.w3.org/2018/credentials/v1',
+              'https://w3id.org/vdl/v1',
+              'https://w3id.org/vdl/aamva/v1'
+            ],
+            type: [
+              'Iso18013DriversLicenseCredential'
+            ]
+          },
+          acceptedEnvelopes: ['application/jwt']
+        }
       });
       expect(presentation_definition.constraints.fields[0].path).to.eql([
         '$[\'@context\']',